fix: space에 함께 있던 role을 모듈로 분리 및 api 테스트 완료

+++ src/space/space.service.ts
@@ -1,11 +1,10 @@
 import { Repository } from 'typeorm';
 import { SpaceModel } from './entities/space.entity';
-import { Injectable } from '@nestjs/common';
+import { BadRequestException, Injectable } from '@nestjs/common';
 import { InjectRepository } from '@nestjs/typeorm';
-import { SpaceRoleModel } from './entities/space-role.entity';
-import { CreateSpaceDto } from './dto/create-space.dto';
 import { UserService } from 'src/user/user.service';
-import { CreateSpaceRoleDto } from './dto/create-space-role.dto';
+import { RoleService } from './role/role.service';
+import { CreateSpaceDto } from './dto/create-space.dto';
 
 @Injectable()
 export class SpaceService {
@@ -13,30 +12,20 @@ export class SpaceService {
         @InjectRepository(SpaceModel)
         private readonly spaceRepository: Repository<SpaceModel>,
 
-        @InjectRepository(SpaceRoleModel)
-        private readonly spaceRoleRepository: Repository<SpaceRoleModel>,
-
         private readonly userService: UserService,
+
+        private readonly roleService: RoleService,
     ) { }
 
     async getAllSpaces() {
         const spaces = await this.spaceRepository.find({
-            relations: ['owner', 'spaceRoles', 'participatingUsers']
+            relations: ['owner', 'roles', 'participatingUsers']
         });
         return spaces
     }
 
-    generateEntryCode() {
-        const characters = 'ABCDEFGHIJKLMNOPQRSTUVWXYZabcdefghijklmnopqrstuvwxyz0123456789';
-        let entryCode = '';
-        for (let i = 0; i < 8; i++) {
-            entryCode += characters.charAt(Math.floor(Math.random() * characters.length));
-        }
-        return entryCode;
-    }
-
     async createSpace(ownerId: number, spaceDto: CreateSpaceDto) {
-        const { spaceName, spaceLogo, spaceRoles } = spaceDto;
+        const { spaceName, spaceLogo, roles } = spaceDto;
         const newSpace = await this.spaceRepository.save({
             ownerId,
             spaceName,
@@ -45,31 +34,21 @@ export class SpaceService {
 
         const spaceId = newSpace.id;
 
-        const ownerEntryCode = this.generateEntryCode();
-        await this.spaceRoleRepository.save({
-            spaceId, role: '소유자', isAdministrator: true, entryCode: ownerEntryCode
-        });
+        const ownerRole = await this.roleService.createDefaultRole(spaceId);        
+        await this.addUserToSpace( spaceId, ownerId, ownerRole.entryCode);
 
-        for (const spaceRoleDto of spaceRoles) {
-            const { role, isAdministrator } = spaceRoleDto;
-            const entryCode = this.generateEntryCode();
-            const spaceRoleObject = await this.spaceRoleRepository.create({
-                spaceId, role, isAdministrator, entryCode
-            });
-            await this.spaceRoleRepository.save(spaceRoleObject);
+        for (const roleDto of roles) {
+            await this.roleService.createNewRole(spaceId, ownerId, roleDto)
         };
 
-        const space = this.addUserToSpace( spaceId, ownerId, ownerEntryCode);
+        const space = this.spaceRepository.findOne({
+            where: {
+                id: spaceId,
+            },
+            relations: ['roles', 'participatingUsers']
+        });
 
         return space;
-    }
-
-    async getSpaceRole(spaceId: number) {
-        return this.spaceRoleRepository.find({
-            where: {
-                spaceId,
-            }
-        });
     }
 
     async addUserToSpace(spaceId: number, userId: number, entryCode: string) {
@@ -77,33 +56,33 @@ export class SpaceService {
             where: {
                 id: spaceId
             },
-            relations: ['spaceRoles', 'participatingUsers']
+            relations: ['roles', 'participatingUsers']
         });
 
         if (!space) {
-            throw new Error('Space가 없습니다.');
+            throw new BadRequestException('존재하지 않는 Space입니다.');
         }
 
-        const spaceRoles = space.spaceRoles;
-        const spaceRole = spaceRoles.find((spaceRole) => {
-            return spaceRole.entryCode == entryCode;
+        const roles = space.roles;
+        const role = roles.find((role) => {
+            return role.entryCode == entryCode;
         });
 
-        if (!spaceRole) {
-            throw new Error('입장 코드가 틀렸습니다.')
+        if (!role) {
+            throw new BadRequestException('입장 코드가 틀렸습니다.')
         }
 
-        const bridge = await this.userService.createBridge(spaceId, userId, spaceRole);
+        const bridge = await this.userService.createBridge(spaceId, userId, role);
 
         if (!bridge) {
-            throw new Error('이미 존재하는 유저입니다.');
+            throw new BadRequestException('이미 존재하는 유저입니다.');
         }
 
         const updatedSpace = await this.spaceRepository.findOne({
             where: {
                 id: spaceId
             },
-            relations: ['spaceRoles', 'participatingUsers']
+            relations: ['roles', 'participatingUsers']
         });
         return updatedSpace;
     }
@@ -113,20 +92,20 @@ export class SpaceService {
             where: {
                 id: spaceId
             },
-            relations: ['spaceRoles', 'participatingUsers']
+            relations: ['roles', 'participatingUsers']
         });
 
         if (!space) {
-            throw new Error('Space가 없습니다.');
+            throw new BadRequestException('존재하지 않는 Space입니다.');
         }
 
         const deleteResult = await this.userService.deleteBridge(spaceId, userId);
 
         if (!deleteResult) {
-            throw new Error('삭제에 실패했습니다');
+            throw new BadRequestException('존재하지 않는 User입니다.');
         }
 
-        return true;
+        return deleteResult;
     }
 
     async deleteSpace(spaceId:number, userId:number) {
@@ -134,55 +113,17 @@ export class SpaceService {
             where: {
                 id: spaceId
             },
-            relations: ['spaceRoles', 'participatingUsers']
+            relations: ['roles', 'participatingUsers']
         });
+
+        if (!space) {
+            throw new BadRequestException('존재하지 않는 Space입니다.');
+        }
 
         if (space.ownerId != userId) {
             throw new Error('Space 소유자가 아닙니다.');
         }
 
         return await this.spaceRepository.delete({ id: spaceId });
-    }
-
-    async createSpaceRole(spaceId: number, userId: number, spaceRoleDto: CreateSpaceRoleDto) {
-        const { role, isAdministrator } = spaceRoleDto;
-
-        const userRole = await this.userService.getRoleOfUser(spaceId, userId);
-
-        if (!userRole) {
-            throw new Error('Space에 없는 User입니다.')
-        }
-
-        if (!userRole.isAdministrator) {
-            throw new Error('관리자가 아닙니다.');
-        }
-
-        const entryCode = this.generateEntryCode();
-        const spaceRoleObject = await this.spaceRoleRepository.create({
-            spaceId, role, isAdministrator, entryCode
-        });
-
-        const existingRole = await this.spaceRoleRepository.exists({
-            where: {
-                spaceId,
-                role
-            }
-        });
-        if (existingRole) {
-            throw new Error('이미 존재하는 이름의 역할입니다.');
-        }
-
-        const newRole = await this.spaceRoleRepository.save(spaceRoleObject);
-        return newRole;
-    }
-
-    async deleteSpaceRole(spaceId:number, userId: number, roleId: number) {
-        const userRole = await this.userService.getRoleOfUser(spaceId, userId)
-
-        if (!userRole.isAdministrator) {
-            throw new Error('관리자가 아닙니다.');
-        }
-
-        return await this.spaceRoleRepository.delete({id: roleId});
     }
 }
